# infrastructure/api-gateway/openapi.yaml
openapi: "3.0.1"
info:
  title: "Panda CRM API"
  description: "API Gateway for Panda CRM microservices"
  version: "1.0.0"
servers:
  - url: "https://api.pandacrm.com"
    description: "Production"
  - url: "https://api-staging.pandacrm.com"
    description: "Staging"
...
paths:
...
  /api/leads:
    get:
      summary: List leads
      tags: [Leads]
      security:
        - CognitoAuthorizer: []
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/leads"
        type: http_proxy
        httpMethod: GET
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
    post:
      summary: Create lead
      tags: [Leads]
      security:
        - CognitoAuthorizer: []
      x-amazon-apigateway-integration:
        uri: "http://${ALB_DNS}/api/leads"
        type: http_proxy
        httpMethod: POST
        connectionType: VPC_LINK
        connectionId: "${VPC_LINK_ID}"
...
components:
  securitySchemes:
    CognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        type: cognito_user_pools
        providerARNs:
          - "arn:aws:cognito-idp:us-east-2:679128292059:userpool/us-east-2_e02zbxuZ2"